+++ src/test/java/study/SetCollectionTest.java
@@ -1,4 +1,4 @@
-package step1;
+package study;
 
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
