+++ src/test/java/study/StringClassTest.java
@@ -1,4 +1,4 @@
-package step1;
+package study;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
